% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-annotation.R
\name{analyse.annotation}
\alias{analyse.annotation}
\title{A function to summarize the presence of some genes based on annotation}
\usage{
analyse.annotation(bacteria.table, annotation.list, collicin)
}
\arguments{
\item{collicin}{}
}
\value{

}
\description{
A function to summarize the presence of some genes based on annotation
}
