% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.consensus.sequence.R
\name{compute.consensus.sequence}
\alias{compute.consensus.sequence}
\title{A function to compute consensus sequence for each gene}
\usage{
compute.consensus.sequence(species, collicin, genDir, outDir)
}
\arguments{
\item{outDir}{}
}
\value{

}
\description{
A function to compute consensus sequence for each gene
}
