% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-genes-blast-based.R
\name{exctratblast}
\alias{exctratblast}
\title{Title}
\usage{
exctratblast(
  accession.list,
  repertoire_genome,
  repertoire_gene_annotation,
  repertoire_blast_based
)
}
\arguments{
\item{repertoire_blast_based}{}
}
\value{

}
\description{
Title
}
