% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.all.genes.annotionbased.R
\name{extract.all.genes.annotationbased}
\alias{extract.all.genes.annotationbased}
\title{Title}
\usage{
extract.all.genes.annotationbased(
  species,
  genemlst,
  annotationDir,
  genomeDir,
  outDir
)
}
\value{

}
\description{
Title
}
