% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.closest.R
\name{extract.closest}
\alias{extract.closest}
\title{Title}
\usage{
extract.closest(
  genomePath,
  genepath,
  lengthconf = 95,
  identconf = 95,
  offset = 0
)
}
\arguments{
\item{offset}{}
}
\value{

}
\description{
Title
}
