% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract1gene.R
\name{extract1gene}
\alias{extract1gene}
\title{Title}
\usage{
extract1gene(
  selectedspecies = "ecoli",
  selectedgene = "tolB",
  repertoire_gene,
  repertoire_annotation,
  repertoire_genome
)
}
\arguments{
\item{repertoire_genome}{}
}
\value{

}
\description{
Title
}
