% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_not_repicated.R
\name{gene_not_duplicated}
\alias{gene_not_duplicated}
\title{Title A function to exclude replicated genes within a species}
\usage{
gene_not_duplicated(species, geneDir, outDir, collicin)
}
\arguments{
\item{collicin}{}
}
\value{

}
\description{
Title A function to exclude replicated genes within a species
}
