% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.from.consensus.R
\name{phylo.from.consensus}
\alias{phylo.from.consensus}
\title{A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory}
\usage{
phylo.from.consensus(consensusDir, bacteria.table)
}
\arguments{
\item{consensusDir}{the directory with the consensus}
}
\value{

}
\description{
A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory
}
