% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.from.consensus.R
\name{plot.phylo.from.consensus}
\alias{plot.phylo.from.consensus}
\title{A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory}
\usage{
\method{plot}{phylo.from.consensus}(consensusDir)
}
\arguments{
\item{consensusDir}{the directory with the consensus}
}
\value{

}
\description{
A function to plot phylogenetic tree of all genes based on fasta file in the consensus directory
}
