% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writteextract1gene.R
\name{writteextract1gene}
\alias{writteextract1gene}
\title{Title}
\usage{
writteextract1gene(
  accession.list,
  collicin,
  repertoire__gene_annotation,
  repertoire_annotation,
  repertoire_genome
)
}
\arguments{
\item{repertoire__gene_annotation}{}
}
\value{

}
\description{
Title
}
